Move UI mocks into separate file
