+++ frontend/test/metabase/visualizations/components/LineAreaBarRenderer.tz.unit.spec.js
@@ -1,4 +1,4 @@
-import "__support__/mocks"; // included explicitly whereas with integrated tests it comes with __support__/integrated_tests
+import "__support__/ui-mocks"; // included explicitly whereas with integrated tests it comes with __support__/integrated_tests
 import testAcrossTimezones from "__support__/timezones";
 
 import _ from "underscore";
